refactor: move io.py under utils folder
