+++ pinterest_dl/low_level/webdriver/browser.py
@@ -8,7 +8,7 @@ from selenium.webdriver.remote.webdriver import WebDriver
 
 from pinterest_dl.data_model.browser_version import BrowserVersion
 from pinterest_dl.low_level.webdriver.driver_installer import ChromeDriverInstaller
-from pinterest_dl.low_level.ops import io
+from pinterest_dl.utils import io
 
 
 class Browser:
