+++ pinterest_dl/scrapers/scraper_webdriver.py
@@ -5,9 +5,9 @@ from typing import Any, List, Literal, Optional, Tuple, Union
 
 from selenium.webdriver.remote.webdriver import WebDriver
 
-from pinterest_dl.low_level.ops import io
 from pinterest_dl.low_level.webdriver.browser import Browser
 from pinterest_dl.low_level.webdriver.pinterest_driver import PinterestDriver, PinterestImage
+from pinterest_dl.utils import io
 
 from .scraper_base import _ScraperBase
